+++ CHANGELOG.md
@@ -4,10 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.1.1] - 2019-12-29
+## [1.2.0] - 2019-01-02
 
 ### Changed
-* bump `react-button-a11y`
+* fix commonjs bundle
+* bump button lib
+* higher react peer dependency (from 16.3 to 16.8)
 
 ## [1.1.0] - 2019-12-29